fix security problem and refactor

+++ controller/ControllerBoard.php
@@ -26,7 +26,7 @@ class ControllerBoard extends Controller {
         $errors = [];
         $user = $this->get_user_or_redirect();
         $board = $this::get_board_if_exist();
-        if($board && $user->has_permission($board->get_board_id())){
+        if($board && $user->has_permission_aac($board->get_board_id())){
             (new View("board"))->show(array("board" => $board, "user" => $user, "errors" => $errors));
         }else{
             $this->redirect("board","index");
@@ -37,34 +37,36 @@ class ControllerBoard extends Controller {
         $errors = [];
         $user = $this->get_user_or_redirect();
         if (isset($_GET["param1"]) && $_GET["param1"] !== "") {
-
             $board = Board::get_board($_GET["param1"]);
-
-            if($user != $board->get_author() && !$user->is_admin()) {
-                $errors[] = "You cannot delete someone else's board";
-                (new View("board"))->show(array("board" => $board, "user" => $user, "errors" => $errors));
-            } else if($board->has_columns()) { // s'il y a des colonnes il faut confirmation
-                if(!(isset($_POST['confirmation']) && $_POST['confirmation'])) {
-                    $this->redirect("board", "delete_confirm", $board->get_board_id()); // pas de confirmation -> redirection
-                }
-                else {
-                    $board = $this->delete_board(); // confirmation -> on delete
-    
+            if($board && $user->has_permission_aac($board->get_board_id())){
+                if($user != $board->get_author() && !$user->is_admin()) {
+                    $errors[] = "You cannot delete someone else's board";
+                    (new View("board"))->show(array("board" => $board, "user" => $user, "errors" => $errors));
+                } else if($board->has_columns()) { // s'il y a des colonnes il faut confirmation
+                    if(!(isset($_POST['confirmation']) && $_POST['confirmation'])) {
+                        $this->redirect("board", "delete_confirm", $board->get_board_id()); // pas de confirmation -> redirection
+                    }
+                    else {
+                        $board = $this->delete_board(); // confirmation -> on delete
+        
+                        if ($board) {
+                            $this->redirect("board", "index");
+                        } else {
+                            throw new Exception("Wrong/missing ID or action not permitted");
+                        }
+                    } 
+                } else {   
+                    $_POST['board_id'] = $board->get_board_id();
+                    $board = $this->delete_board();
+        
                     if ($board) {
                         $this->redirect("board", "index");
                     } else {
                         throw new Exception("Wrong/missing ID or action not permitted");
                     }
-                } 
-            } else {   
-                $_POST['board_id'] = $board->get_board_id();
-                $board = $this->delete_board();
-    
-                if ($board) {
-                    $this->redirect("board", "index");
-                } else {
-                    throw new Exception("Wrong/missing ID or action not permitted");
                 }
+            }else {    
+                $this->redirect("board", "index");
             }
         } else {    
             $this->redirect("board", "index");
@@ -75,7 +77,7 @@ class ControllerBoard extends Controller {
     public function delete_confirm() {
         $user = $this->get_user_or_redirect();
         $board = $this::get_board_if_exist();
-        if($board && $user->has_permission($board->get_board_id())){
+        if($board && $user->has_permission_aac($board->get_board_id())){
             (new View("board_delete"))->show(array("user" => $user, "board" => $board));
         }else{
             $this->redirect("board","index");
@@ -84,11 +86,10 @@ class ControllerBoard extends Controller {
 
     private function delete_board() {
         $user = $this->get_user_or_redirect();
-
         if (isset($_POST['board_id']) && $_POST['board_id'] != "") {
             $board_id = $_POST['board_id'];
             $board = Board::get_board($board_id);
-            if ($board) {
+            if ($board && $user->has_permission_aac($board->get_board_id())) {
                 return $board->delete($user);
             } 
         }
@@ -130,7 +131,7 @@ class ControllerBoard extends Controller {
         $user = $this->get_user_or_redirect();
         $board = $this::get_board_if_exist();
         $errors = [];
-        if($board && $user->has_permission($board->get_board_id())){
+        if($board && $user->has_permission_aac($board->get_board_id())){
             (new View("board_edit"))->show(array("board" => $board, "user" => $user, "errors" => $errors));
         }else{
             $this->redirect("board","index");
@@ -143,13 +144,19 @@ class ControllerBoard extends Controller {
         if (isset($_POST['board_id']) && isset($_POST['title'])) {
             $proposed_title = $_POST["title"];
             $board = Board::get_board($_POST['board_id']);
-            $errors = $board->validate_board_name($proposed_title);
-            if (count($errors) == 0) { 
-                $board->set_title($_POST["title"]);
-                $board->update(); 
-                $this->redirect("board", "board", $board->get_board_id());
+            if($board && $user->has_permission_aac($board->get_board_id())){
+                $errors = $board->validate_board_name($proposed_title);
+                if (count($errors) == 0 ) { 
+                    $board->set_title($_POST["title"]);
+                    $board->update(); 
+                    $this->redirect("board", "board", $board->get_board_id());
+                }
+                (new View("board_edit"))->show(array("board" => $board, "user" => $user, "errors" => $errors, "proposed_title" => $proposed_title));
+            }else{
+                $this->redirect("board","index");
             }
-            (new View("board_edit"))->show(array("board" => $board, "user" => $user, "errors" => $errors, "proposed_title" => $proposed_title));
+        }else{
+            $this->redirect("board","index");
         }
     }
 
@@ -157,8 +164,7 @@ class ControllerBoard extends Controller {
         $user = $this->get_user_or_redirect();
         $board = $this::get_board_if_exist();
         $errors = [];
-        //attention ici, seul l'admin ou l'auteur du tableau peut voir les collaborateurs
-        if($board && ($user->is_admin() || $board->get_author_id() === $user->get_user_id())){
+        if($board && ($user->has_permission_aa($board->get_author_id()))){
             (new View("board_collaborators"))->show(array("board" => $board, "user" => $user, "errors" => $errors));
         }else{
             $this->redirect("board","index");
@@ -167,15 +173,20 @@ class ControllerBoard extends Controller {
 
     public function add_collaborator(){
         $errors = [];
+        $user = $this->get_user_or_redirect();
         if (isset($_POST['board_id']) && isset($_POST['new_collaborator_id'])) {
             $new_collaborator_id = $_POST["new_collaborator_id"];
             $board = Board::get_board($_POST['board_id']);
-            $errors = $board->validate_board_new_collaborator($new_collaborator_id);
-            if (count($errors) == 0) { 
-                $board->add_new_collaborator($new_collaborator_id); 
-                $this->redirect("board", "collaborators", $board->get_board_id());
+            if ($board && $user->has_permission_aa($board->get_author_id())) { 
+                $errors = $board->validate_board_new_collaborator($new_collaborator_id);
+                if (count($errors) == 0) { 
+                    $board->add_new_collaborator($new_collaborator_id); 
+                    $this->redirect("board", "collaborators", $board->get_board_id());
+                }
+                (new View("board_collaborators"))->show(array("board" => $board, "user" => $user, "errors" => $errors));
+            }else{
+                $this->redirect("board","index");
             }
-            (new View("board_collaborators"))->show(array("board" => $board, "user" => $user, "errors" => $errors));
         }else{
             $this->redirect("board","index");
         }
@@ -184,12 +195,13 @@ class ControllerBoard extends Controller {
     public function remove_collaboration_confirm() {
         $user = $this->get_user_or_redirect();
         if (isset($_GET["param1"]) && $_GET["param1"] !== "" && isset($_GET["param2"]) && $_GET["param2"] !== "") {
-        $board_id = $_GET["param1"];
-        $collaborator_id = $_GET["param2"];
+            $board_id = $_GET["param1"];
+            $collaborator_id = $_GET["param2"];
         }else{
             $this->redirect("board","index");
         }
-        if($user->has_permission($board_id)){
+        $board = Board::get_board($board_id);
+        if($board && $user->has_permission_aa($board->get_author_id())){
             (new View("collaboration_delete"))->show(array("user" => $user, "board_id" => $board_id, "collaborator_id" => $collaborator_id));
         }else{
             $this->redirect("board","index");
@@ -197,15 +209,18 @@ class ControllerBoard extends Controller {
     }
 
     public function collaborator_delete(){
+        $user = $this->get_user_or_redirect();
         $errors = [];
         if (isset($_POST['board_id']) && isset($_POST['collaborator_id'])) {
             $collaborator_id = $_POST["collaborator_id"];
             $board = Board::get_board($_POST['board_id']);
-            if (count($errors) == 0) { 
+            if ($board && $user->has_permission_aa($board->get_author_id())) { 
                 $board->remove_collaborator($collaborator_id); 
                 $this->redirect("board", "collaborators", $board->get_board_id());
+                (new View("board_collaborators"))->show(array("board" => $board, "user" => $user, "errors" => $errors));
+            }else{
+                $this->redirect("board","index");
             }
-            (new View("board_collaborators"))->show(array("board" => $board, "user" => $user, "errors" => $errors));
         }else{
             $this->redirect("board","index");
         }
@@ -219,7 +234,7 @@ class ControllerBoard extends Controller {
                 $registered_board_id = ($_POST["board_id"]);
             }
             $board = Board::get_board_by_title($_POST["board_title"]);
-            if($registered_board_id===0 && $board){
+            if($board && $registered_board_id===0){
                 $res = "false";
             }else if($board && $registered_board_id !== $board->get_board_id() ){
                 $res = "false";
